Node updating progress, key handling start
I've broken the d3.json... data loading method up so that the main work
is done in a function called start(), that can be called after updating
the nodes/links structures by going force.nodes(newThing). etc. The easiest
way to do this seems to be to get the existing structure with a no-argument
call (force.nodes()), then append to it, then call start(). It seems to work.