Add Links complete, working on remove links
The links now add through to the database. Did this with ajax calls
to django, just like the more sophisticated quant browsing functions.
There is some CSRF trickery that needs to be done in the ajax call,
but sticking the {{csrf_token}} in the data of the ajax POST call
is all that's needed, despite the sophisticated crap you find all
over the internets.
I had to reconfigure the way you calculate the link indices for d3,
as i'd been sorting the node list, then using the node ids, forgetting
that there can be discontinuities in the node indices, causing their
position in the list to no longer match their index. Fixed with a
small function, but oops haven't amended the comment yet.