Add and Edit Nodes. Crazy Typing System.
Can now add and edit nodes using 'a' and 'e'.
Nodes search for templates (also in nodes) by using "linking to template"
specific edges, and if they don't have them they explore backwards (breadth first)
to find one that does, keeping a stack of visits to detect loops, in which case no
template is returned. The node and its template then render (add, edit or view) to
a div that can be placed.
Adding a nod prompts the user for a name, and creates a new node with that name, no
content, and the type specified by the overhead drop down box.