Much awesome. So win.
Can delete nodes with 'd' key, and add at mouse position, even after
panning and zooming.
Delete is not turned on yet at the database level.
Needed a key function in the data binding for nodes (still don't understand
how this works), but without it deleting a node was messing everything up.