Random colors and sizes. Sliders for graph params.
Randomly generated sizes and colors from the django to sample how this is done.
We've got a bunch of jquery sliders that control such things as gravity, charge,
distances, theta... So we don't have to use the mouse for this anymore, which is
good because holding down shift and mousewheeling on top of this thing is somehow
disabling the keyboard.